refactor: SkateCastComponent
