+++ src/app/skatecast/page.tsx
@@ -1,171 +1,39 @@
 "use client"
-import { MarkdownRenderers } from "@/app/upload/utils/MarkdownRenderers"
+import { VStack } from "@chakra-ui/react"
 import { useHiveUser } from "@/contexts/UserContext"
 import { useComments } from "@/hooks/comments"
-import { vote } from "@/lib/hive/client-functions"
-import {
-  formatDate,
-  transformIPFSContent,
-  transformShortYoutubeLinksinIframes,
-} from "@/lib/utils"
-import {
-  Avatar,
-  Box,
-  Button,
-  Divider,
-  Flex,
-  HStack,
-  Text,
-  Textarea,
-  VStack,
-  Input,
-  Badge,
-  Link,
-} from "@chakra-ui/react"
-import * as dhive from "@hiveio/dhive"
-import { useEffect, useMemo, useRef, useState } from "react"
-import { useDropzone } from "react-dropzone"
-import { FaDollarSign, FaImage, FaRegComment, FaRegHeart, FaHeart } from "react-icons/fa"
-import InfiniteScroll from "react-infinite-scroll-component"
-import ReactMarkdown from "react-markdown"
-import { BeatLoader } from "react-spinners"
-import rehypeRaw from "rehype-raw"
-import remarkGfm from "remark-gfm"
-import { uploadFileToIPFS } from "../upload/utils/uploadToIPFS"
-import { commentWithPrivateKey } from "@/lib/hive/server-functions"
-import { useCasts } from "@/hooks/casts"
-import TipButton from "@/components/PostCard/TipButton"
+import AvatarList from "./AvatarList"
+import PostBox from "./PostBox"
 import LoadingComponent from "./loadingComponent"
 import AvatarMediaModal from "./mediaModal"
-import { handleVote } from "./utils/handleFeedVote"
-import { useReward } from "react-rewards"
-import { comment } from "@uiw/react-md-editor"
-import { voting_value2 } from "@/components/PostCard/calculateHiveVotingValueForHiveUser"
+import {  vote } from "@/lib/hive/client-functions"
+import * as dhive from "@hiveio/dhive"
+import CommentList from "./CommentsList"
+import { commentWithPrivateKey } from "@/lib/hive/server-functions"
+import { useState, useEffect, useMemo } from "react"
 
 const parent_author = "skatehacker"
 const parent_permlink = "test-advance-mode-post"
-
-const VotingButton = ({ comment, username }: { comment: any, username: any }) => {
-  const initialIsVoted = comment.active_votes?.some((vote: any) => vote.voter === username);
-  const [isVoted, setIsVoted] = useState(initialIsVoted);
-  const [voteCount, setVoteCount] = useState(comment.active_votes?.length || 0);
-  const rewardId = `reward-${comment.id}`;
-  const { reward } = useReward(rewardId, "emoji", {
-    emoji: ["$", "*", "#"],
-    spread: 60,
-  });
-
-  const handleVoteClick = async () => {
-    const newIsVoted = !isVoted;
-    await handleVote(comment.author, comment.permlink, username ?? "");
-    setIsVoted(newIsVoted);
-
-    setVoteCount((prevVoteCount: number) => newIsVoted ? prevVoteCount + 1 : prevVoteCount - 1);
-
-    if (newIsVoted) {
-      reward();
-    }
-
-  };
-
-  return (
-    <Button
-      onClick={handleVoteClick}
-      colorScheme="green"
-      variant="ghost"
-      leftIcon={isVoted ? <FaHeart /> : <FaRegHeart />}
-    >
-      <span
-        id={rewardId}
-        style={{
-          position: "absolute",
-          left: "50%",
-          bottom: "15px",
-          transform: "translateX(-50%)",
-          zIndex: 5,
-        }}
-      />
-      {voteCount}
-    </Button>
-  );
-};
-
 
 const SkateCast = () => {
   const { comments, addComment, isLoading } = useComments(
     parent_author,
     parent_permlink
   )
-  const [visiblePosts, setVisiblePosts] = useState(40)
+  const [visiblePosts, setVisiblePosts] = useState(20)
   const [postBody, setPostBody] = useState("")
   const reversedComments = comments?.slice().reverse()
   const user = useHiveUser()
   const username = user?.hiveUser?.name
   const [mediaModalOpen, setMediaModalOpen] = useState(false)
   const [media, setMedia] = useState<string[]>([])
-  const [mediaComments, setMediaComments] = useState(new Set())
+  const [mediaComments, setMediaComments] = useState<Set<number>>(new Set())
   const [mediaDictionary, setMediaDictionary] = useState(new Map())
-  const PINATA_GATEWAY_TOKEN = process.env.NEXT_PUBLIC_PINATA_GATEWAY_TOKEN
-  const [isUploading, setIsUploading] = useState(false)
-  const inputRef = useRef<HTMLInputElement>(null)
   const [hasPosted, setHasPosted] = useState(false)
-  const rewardId = comments?.[0]?.id ? "postReward" + comments[0].id : ""
-  const [userVotingValue, setUserVotingValue] = useState(0)
-  const userHasVoted = (username: string) => {
-
-    return comments?.some((comment) =>
-      comment.active_votes?.some((vote) => vote.voter === username)
-    )
-
-  }
-
-  useEffect(() => {
-    const getVotingValue = async () => {
-      try {
-        if (!username) return;
-        const vote_value = await voting_value2(username);
-        setUserVotingValue(Number(vote_value.toFixed(2)));
-      } catch (error) {
-        console.error("Failed to calculate voting value:", error);
-      }
-    }
-    if (username) {
-      getVotingValue();
-    }
-  }, [username]);
-
-
-
-
-
-  const { getRootProps, getInputProps } = useDropzone({
-    noClick: true,
-    noKeyboard: true,
-    onDrop: async (acceptedFiles) => {
-      setIsUploading(true)
-      for (const file of acceptedFiles) {
-        const ipfsData = await uploadFileToIPFS(file) // Use the returned data directly
-        if (ipfsData !== undefined) {
-          // Ensure ipfsData is not undefined
-          const ipfsUrl = `https://ipfs.skatehive.app/ipfs/${ipfsData.IpfsHash}?pinataGatewayToken=${PINATA_GATEWAY_TOKEN}`
-          const markdownLink = file.type.startsWith("video/")
-            ? `<iframe src="${ipfsUrl}" allowfullscreen></iframe>`
-            : `![Image](${ipfsUrl})`
-          setPostBody((prevMarkdown) => `${prevMarkdown}\n${markdownLink}\n`)
-        }
-      }
-      setIsUploading(false)
-    },
-    accept: {
-      "image/*": [".png", ".gif", ".jpeg", ".jpg"],
-      "video/*": [".mp4", ".mov"],
-    },
-    multiple: false,
-  })
 
   useEffect(() => {
     if (comments) {
-      const mediaSet = new Set()
+      const mediaSet = new Set<number>();
       const mediaDict = new Map()
       comments?.forEach((comment) => {
         const media = comment.body.match(
@@ -176,7 +44,9 @@ const SkateCast = () => {
             ? "video"
             : "image"
         if (media) {
-          mediaSet.add(comment.id)
+          if (comment.id !== undefined) {
+            mediaSet.add(comment.id);
+          }
           mediaDict.set(comment.id, { media, type: mediaType })
         }
       })
@@ -204,14 +74,12 @@ const SkateCast = () => {
   }, [comments, mediaComments])
 
   const handlePost = async () => {
-    console.log("handlePost")
     const permlink = new Date()
       .toISOString()
       .replace(/[^a-zA-Z0-9]/g, "")
       .toLowerCase()
-    console.log("permlink", permlink)
+
     const loginMethod = localStorage.getItem("LoginMethod")
-    console.log("loginMethod", loginMethod)
 
     if (!username) {
       console.error("Username is missing")
@@ -256,7 +124,6 @@ const SkateCast = () => {
 
           if (response.success) {
             setPostBody("")
-            console.log("Comment posted successfully")
             addComment(postData)
           }
         } catch (error) {
@@ -321,6 +188,19 @@ const SkateCast = () => {
     }
   }
 
+  const handleVote = async (author: string, permlink: string) => {
+    if (!username) {
+      console.error("Username is missing")
+      return
+    }
+    vote({
+      username: username,
+      permlink: permlink,
+      author: author,
+      weight: 10000,
+    })
+  }
+
   const handleMediaAvatarClick = (commentId: number) => {
     const media = mediaDictionary.get(commentId)
     console.log("media", media)
@@ -351,18 +231,6 @@ const SkateCast = () => {
     }
   }
 
-  const handleImageUploadClick = () => {
-    console.log("inputRef", inputRef)
-    if (inputRef.current) {
-      inputRef.current.click()
-    }
-  }
-
-  function textAreaAdjust(element: any) {
-    element.style.height = "1px"
-    element.style.height = 25 + element.scrollHeight + "px"
-  }
-
   return isLoading ? (
     <LoadingComponent />
   ) : (
@@ -378,177 +246,28 @@ const SkateCast = () => {
         onClose={() => setMediaModalOpen(false)}
         media={media}
       />
-      <HStack
-        flexWrap={"nowrap"}
-        w={"100%"}
-        css={{ "&::-webkit-scrollbar": { display: "none" } }}
-        overflowX="auto"
-        minHeight={"60px"}
-        px={4}
-      >
-        {sortedComments?.map((comment, index, commentsArray) => {
-          const isDuplicate =
-            commentsArray.findIndex((c) => c.author === comment.author) !==
-            index
-          if (isDuplicate) {
-            return null
-          }
-          return (
-            <Avatar
-              key={comment.id}
-              size="md"
-              src={`https://images.ecency.com/webp/u/${comment.author}/avatar/small`}
-              border={
-                mediaComments.has(comment.id) ? "2px solid limegreen" : "none"
-              }
-              cursor={"pointer"}
-              onClick={() => handleMediaAvatarClick(Number(comment.id))}
-            />
-          )
-        })}
-        <Divider />
-      </HStack>
+      <AvatarList
+        sortedComments={sortedComments}
+        mediaComments={mediaComments}
+        handleMediaAvatarClick={handleMediaAvatarClick}
+      />
       {user.hiveUser !== null && (
-        <Box p={4} width={"100%"} bg="black" color="white" {...getRootProps()}>
-          <Flex>
-            <Avatar
-              borderRadius={10}
-              boxSize={12}
-              src={`https://images.ecency.com/webp/u/${username}/avatar/small`}
-            />
-            <Textarea
-              border="none"
-              _focus={{
-                border: "none",
-                boxShadow: "none",
-              }}
-              placeholder="What's happening?"
-              onChange={(e) => setPostBody(e.target.value)}
-              value={postBody}
-              overflow={"hidden"}
-              resize={"vertical"}
-              onKeyUp={(e) => textAreaAdjust(e.target)}
-            />
-          </Flex>
-          <HStack>
-            {postBody.includes("![Image](") && (
-              <Box>
-                <img
-                  src={postBody.match(/!\[Image\]\((.*?)\)/)?.[1]}
-                  alt="markdown-image"
-                  width="100%"
-                />
-              </Box>
-            )}
-            {postBody.includes("<iframe") && (
-              <Box>
-                <video
-                  src={
-                    postBody.match(
-                      /<iframe src="(.*?)" allowfullscreen><\/iframe>/
-                    )?.[1]
-                  }
-                  controls
-                  muted
-                  width="100%"
-                />
-              </Box>
-            )}
-          </HStack>
-          <HStack justifyContent="space-between" m={4}>
-            <Input
-              id="md-image-upload"
-              type="file"
-              style={{ display: "none" }}
-              {...getInputProps({ refKey: "ref" })}
-              ref={inputRef}
-            />
-            <FaImage
-              color="#ABE4B8"
-              cursor="pointer"
-              onClick={handleImageUploadClick}
-            />
-            <Button
-              colorScheme="green"
-              variant="outline"
-              ml="auto"
-              onClick={handlePost}
-              isLoading={isUploading}
-            >
-              Post
-            </Button>
-          </HStack>
-          <Divider mt={4} />
-        </Box>
+        <PostBox
+          username={username}
+          postBody={postBody}
+          setPostBody={setPostBody}
+          handlePost={handlePost}
+        />
       )}
-      <Box width={"full"} >
-        <InfiniteScroll
-          dataLength={visiblePosts}
-          next={() => setVisiblePosts(visiblePosts + 3)}
-          hasMore={visiblePosts < (comments?.length ?? 0)}
-          loader={
-            <Flex justify="center">
-              <BeatLoader size={8} color="darkgrey" />
-            </Flex>
-          }
-          style={{ overflow: "hidden" }}
-        >
-          {reversedComments?.slice(0, visiblePosts).map((comment) => (
-            <Box key={comment.id} p={4} width="100%" bg="black" color="white">
-              <Flex>
-                <Link cursor={'pointer'} href={`/profile/${comment.author}`} key={comment.id}>
-                  <Avatar
-                    borderRadius={10}
-                    boxSize={12}
-                    src={`https://images.ecency.com/webp/u/${comment.author}/avatar/small`}
-                  />
-                </Link>
-                <HStack ml={4}>
-                  <Link cursor={'pointer'} href={`/profile/${comment.author}`} key={comment.id}>
-                    <Text fontWeight="bold">{comment.author}</Text>
-                  </Link>
-                  <Text ml={2} color="gray.400">
-                    {formatDate(String(comment.created))}
-                  </Text>
-                  <Badge
-                    variant="ghost"
-                    color={"green.400"}>
-                    <HStack>
-                      <FaDollarSign />
-                      <Text>
-                        {getTotalPayout(comment)} USD
-                      </Text>
-                    </HStack>
-                  </Badge>
-                </HStack>
-              </Flex>
-              <Box ml={"64px"} mt={4}>
-                <ReactMarkdown
-                  components={MarkdownRenderers}
-                  rehypePlugins={[rehypeRaw]}
-                  remarkPlugins={[remarkGfm]}>
-                  {transformIPFSContent(
-                    transformShortYoutubeLinksinIframes(comment.body)
-                  )}
-                </ReactMarkdown>
-              </Box>
-              <Flex ml={12} justifyContent={"space-between"} mt={4}>
-                <Button
-                  colorScheme="green"
-                  variant="ghost"
-                  leftIcon={<FaRegComment />}
-                  onClick={() => handleCommentIconClick(comment)}
-                >
-                  {comment.children}
-                </Button>
-                <VotingButton comment={comment} username={username} />
-                <TipButton author={comment.author} />
-              </Flex>
-              <Divider mt={4} />
-            </Box>
-          ))}
-        </InfiniteScroll>
-      </Box>
+      <CommentList
+        comments={reversedComments}
+        visiblePosts={visiblePosts}
+        setVisiblePosts={setVisiblePosts}
+        username={username}
+        handleCommentIconClick={handleCommentIconClick}
+        handleVote={handleVote}
+        getTotalPayout={getTotalPayout}
+      />
     </VStack>
   )
 }
